fix for PY3.6: added compatibility layer for ChoiceEnum
