+++ shop/models/fields.py
@@ -1,13 +1,11 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
-import enum
-
 from django.conf import settings
 from django.db import models
-from django.utils.six import python_2_unicode_compatible, with_metaclass, string_types
-from django.utils.translation import ugettext_lazy as _, ugettext
+from django.utils.translation import ugettext_lazy as _
 
+from .choice_enum import ChoiceEnum
 
 postgresql_engine_names = [
     'django.db.backends.postgresql',
@@ -29,37 +27,6 @@ class JSONField(_JSONField):
         name, path, args, kwargs = super(JSONField, self).deconstruct()
         del kwargs['default']
         return name, path, args, kwargs
-
-
-class ChoiceEnumMeta(enum.EnumMeta):
-    def __call__(cls, value, *args, **kwargs):
-        if isinstance(value, string_types):
-            try:
-                value = cls.__members__[value]
-            except KeyError:
-                pass  # let the super method complain
-        return super(ChoiceEnumMeta, cls).__call__(value, *args, **kwargs)
-
-
-@python_2_unicode_compatible
-class ChoiceEnum(with_metaclass(ChoiceEnumMeta, enum.Enum)):
-    """
-    Utility class to handle choices in Django model fields
-    """
-    def __str__(self):
-        return ugettext('.'.join((self.__class__.__name__, self.name)))
-
-    @classmethod
-    def default(cls):
-        try:
-            return next(iter(cls))
-        except StopIteration:
-            return None
-
-    @classmethod
-    def choices(cls):
-        choices = [(c.value, str(c)) for c in cls]
-        return choices
 
 
 class ChoiceEnumField(models.PositiveSmallIntegerField):
